+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Includes new 8.0 ffmpeg version bundled with docker and go 1.25 upgrade with new
  - fixed /public/static routes issue
  - shares redirect to login - https://github.com/gtsteffaniak/filebrowser/issues/1109
  - some static assets not available to anonymous user - https://github.com/gtsteffaniak/filebrowser/issues/1102
+ - more safari style issues https://github.com/gtsteffaniak/filebrowser/issues/1110
 
 ## v0.8.1-beta